Extract a functional component for rendering search result rows
Do this to remove a TODO from the codebase.
